BE/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
feedparser
beautifulsoup4
requests
pydantic-settings
google-genai
chromadb
python-dateutil
python-jose[cryptography]
